feat: support role share resource limits from deployment side

+++ src/main/java/com/epam/aidial/cfg/dao/model/ShareResourceLimitEntity.java
@@ -8,5 +8,5 @@ import lombok.Data;
 public class ShareResourceLimitEntity {
 
     private Integer maxAcceptedUsers;
-    private Integer invitationTtl;
+    private Long invitationTtl;
 }
